+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.x branch
 ## 2.25 branch
 ### 2.25.5
+* added `is('ip')` and `is('matchingIp')` request detectors;
 * the abstract `AppController` provided by me-tools has been deprecated and will be removed in a later release. Your
   `AppController` should directly extend `Cake\Controller\Controller`;
 * `AppController` no longer automatically loads any component (previously they were `RequestHandler` and `MeTools.Flash`);